refactor(controller): Added the user id to all express controller functions

+++ backend/controllers/expense.js
@@ -3,8 +3,10 @@ const expenseSchema = require("../models/expenseModel");
 // Method for saving expenses
 exports.addExpense = async (req, res) => {
   const { title, amount, category, description, date } = req.body;
+  const userId = req.auth.userId;
 
   const expense = expenseSchema({
+    userId,
     title,
     amount,
     category,
@@ -35,7 +37,8 @@ exports.addExpense = async (req, res) => {
 // Method for retrieving expenses
 exports.getExpense = async (req, res) => {
   try {
-    const expenses = await expenseSchema.find().sort({ createdAt: -1 }); //making most recent entry appear first
+    const userId = req.auth.userId;
+    const expenses = await expenseSchema.find({userId}).sort({ createdAt: -1 }); //making most recent entry appear first
     res.status(200).json(expenses);
   } catch (error) {
     res.status(500).json({ message: "Server Error" });
@@ -45,9 +48,10 @@ exports.getExpense = async (req, res) => {
 // Deleting entered expenses
 exports.deleteExpense = async (req, res) => {
   const { id } = req.params;
+  const userId = req.auth.userId;
 
   expenseSchema
-    .findByIdAndDelete(id)
+    .findByIdAndDelete({ _id: id, userId })
     .then((expense) => {
       res.status(200).json({ message: "Expense Deleted Successfully" });
     })
